fix(transaccion): correjir el service implementacion al cambio de tipo

+++ src/main/java/ar/edu/unju/fi/arquitecturas/sistemabancario/service/TransaccionService.java
@@ -2,16 +2,17 @@ package ar.edu.unju.fi.arquitecturas.sistemabancario.service;
 
 import ar.edu.unju.fi.arquitecturas.sistemabancario.model.Transaccion;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
 public interface TransaccionService {
 
-    Transaccion realizarDeposito(UUID cuentaId, float monto);
+    Transaccion realizarDeposito(UUID cuentaId, BigDecimal monto);
 
-    Transaccion realizarExtraccion(UUID cuentaId, float monto);
+    Transaccion realizarExtraccion(UUID cuentaId, BigDecimal monto);
 
-    void realizarTransferencia(UUID cuentaOrigenId, UUID cuentaDestinoId, float monto);
+    void realizarTransferencia(UUID cuentaOrigenId, UUID cuentaDestinoId, BigDecimal monto);
 
     List<Transaccion> obtenerHistorialPorCuenta(UUID cuentaId);
